chore: refactor into lib structure
